// app-tablas/src/app/Interfaces/tasas.interface.ts
export interface Tasas{
  dateDesde: Date;
  dateHasta: Date;
  centroMaquina: string;
  subFamilia: string;
  matricula: string;
  bastidor: string;
  tasaTotal: number;
  tasaChasis: number;
  divisa: string;
  implemento1: string;
  tasa1:number;
  divisa1: string;
  implemento2:string;
  tasa2:number;
  divisa2:string;
}